Moving player packages to controller rather than model.
