+++ CardBattleLibrary/src/com/phantomrealm/cardbattle/controller/game/EmptyGameControllerListener.java
@@ -1,8 +1,8 @@
 package com.phantomrealm.cardbattle.controller.game;
 
+import com.phantomrealm.cardbattle.controller.player.PlayerIdentity;
 import com.phantomrealm.cardbattle.model.board.Position;
 import com.phantomrealm.cardbattle.model.card.AttackType;
-import com.phantomrealm.cardbattle.model.player.PlayerIdentity;
 
 /**
  * GameControllerListener which takes no action
