+++ CardBattleLibrary/src/com/phantomrealm/cardbattle/controller/game/GameControllerListener.java
@@ -1,8 +1,8 @@
 package com.phantomrealm.cardbattle.controller.game;
 
+import com.phantomrealm.cardbattle.controller.player.PlayerIdentity;
 import com.phantomrealm.cardbattle.model.board.Position;
 import com.phantomrealm.cardbattle.model.card.AttackType;
-import com.phantomrealm.cardbattle.model.player.PlayerIdentity;
 
 /**
  * Handles callbacks regarding events that occur during the course of gameplay
