+++ CardBattleLibrary/src/com/phantomrealm/cardbattle/controller/player/algorithmic/RandomPlayer.java
@@ -1,25 +1,26 @@
-package com.phantomrealm.cardbattle.model.player.algorithmic;
+package com.phantomrealm.cardbattle.controller.player.algorithmic;
 
 import java.util.List;
+import java.util.Random;
 
+import com.phantomrealm.cardbattle.controller.player.PlayerIdentity;
 import com.phantomrealm.cardbattle.model.board.Board;
 import com.phantomrealm.cardbattle.model.board.Position;
 import com.phantomrealm.cardbattle.model.deck.Deck;
-import com.phantomrealm.cardbattle.model.player.PlayerIdentity;
 
 /**
- * A player who selects the first position of the possible choices for its next move
+ * A player who selects a psuedo-random position of the possible choices for its next move
  * 
  * @author matthewpape
  */
-public class OrderedPlayer extends AlgorithmicPlayer {
+public class RandomPlayer extends AlgorithmicPlayer {
 
 	/**
-	 * Creates an OrderedPlayer with a given identity and deck
+	 * Creates a RandomPlayer with a given identity and deck
 	 * @param identity
 	 * @param deck
 	 */
-	public OrderedPlayer(PlayerIdentity identity, Deck deck) {
+	public RandomPlayer(PlayerIdentity identity, Deck deck) {
 		super(identity, deck);
 	}
 
@@ -29,7 +30,8 @@ public class OrderedPlayer extends AlgorithmicPlayer {
 	@Override
 	protected Position generateMove(Board board) {
 		List<Position> availableMoves = board.getPossibleMoves(getIdentity());
-		return availableMoves.get(0);
+		Random randomGenerator = new Random();
+		return availableMoves.get(randomGenerator.nextInt(availableMoves.size()));
 	}
 
 }
