+++ CardBattleLibrary/src/com/phantomrealm/cardbattle/controller/player/human/ConsoleHumanPlayer.java
@@ -1,16 +1,16 @@
-package com.phantomrealm.cardbattle.model.player.human;
+package com.phantomrealm.cardbattle.controller.player.human;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.List;
 
+import com.phantomrealm.cardbattle.controller.player.BasePlayer;
+import com.phantomrealm.cardbattle.controller.player.PlayerIdentity;
 import com.phantomrealm.cardbattle.model.board.Board;
 import com.phantomrealm.cardbattle.model.board.Position;
 import com.phantomrealm.cardbattle.model.card.Card;
 import com.phantomrealm.cardbattle.model.deck.Deck;
-import com.phantomrealm.cardbattle.model.player.BasePlayer;
-import com.phantomrealm.cardbattle.model.player.PlayerIdentity;
 
 /**
  * A human controlled player who accepts input through the console
