+++ CardBattleLibrary/src/com/phantomrealm/cardbattle/controller/player/minimax/CardCountDisparityPlayer.java
@@ -1,8 +1,8 @@
-package com.phantomrealm.cardbattle.model.player.minimax;
+package com.phantomrealm.cardbattle.controller.player.minimax;
 
+import com.phantomrealm.cardbattle.controller.player.PlayerIdentity;
 import com.phantomrealm.cardbattle.model.board.Board;
 import com.phantomrealm.cardbattle.model.deck.Deck;
-import com.phantomrealm.cardbattle.model.player.PlayerIdentity;
 
 /**
  * A player who favors the move that results in the highest disparity between the number of
